+++ liblitz/draw.c
@@ -94,7 +94,7 @@ static void
 draw_text(Display * dpy, Draw * d)
 {
 	unsigned int x = 0, y = 0, w = 1, h = 1, shortened = 0;
-	size_t len = 0;
+	unsigned int len = 0;
 	static char text[2048];
 
 	if (!d->data)
